Finish past kiai beat before triggering another
This will prevent the ring from being at a static 1.01 scale when there are a lot of notes in a short time.

This shouldn't look jarring at all since in most cases the previous beat should've finished already, and it would be way too fast for the average player to even notice.