Assign break sample to ScorePaddingObject
Not sure if doing it this way is a good either.
But the alternative is that all laned objects have the break sample, and that they play the break sample if hitObject.NeedBreakSample is true. I prefer letting that be determined ahead of time, and to take advantage of the default hitsample flow.