Remove dependence on SimaiSharp for serialisation
SimaiSharp serialisation is *broken by design*. It is full of bugs I can't be bothered to hunt down, and their serialisation is so non-robust it will stroke out if a HitObject is even `float.epsilon` away from where a beat lands, despite it being imperceivable at best.

The precision they work with is un-realistic, and can only be satisfied iff you don't incur *any* floating point error at any point.

Their method of counting beats are also broken in a way that guarantees a beat desynchronisation with hitobjects as soon as any timing changes occur.