Move input handling responsibility back to touch notes
Input handling of touch notes was managed by their hosting playfield.
This was done in order to prevent drag hits from hitting multiple
touches at the same time.

At some point, it was decided that it didn't feel good, and removed that
capability. This meant that TouchPlayfield no longer needs to manage the
inputs.