Fix rounding errors causing incorrect twinning
Due to floating point imprecision, pairs of notes not to be considered twins, even though their differences are minor.

This changes the twinning criteria to be based on time differences rather than naively using absolute time. More specifically, any two notes within 1ms of each other are considered twins.

I am aware that this introduces potential degeneracy when > 2 notes are within 1ms the previous / next, causing all of them to be twinned, even though the maximal difference can be greater than 1ms. I will elect to ignore this, because that is just BS anyways, and it's better than not twinning when it is expected.

HitObjectLines have gotten a similar treatment, but and in such a scenario the line will adopt the *median* start time. Same degeracy scenario applies and ignored.