Don't round any times provided by SimaiSharp
This in theory would've only introduced at most +-0.5ms error to the encoding, simai to compatibility export of lazer. Re-encodes would yield identical timings.

However, in simai, each hitobjects' timing is dependent on the interval from the last hitobject, causing subsequent hitobjects to be shifted too. The quantised encoder then tries to take the error into account, introducing some error in the process, which is then rounded again on the next import. This results in the timing errors compounding with subsequent reimports.